+++ src/main/java/at/hannibal2/skyhanni/features/garden/contest/FarmingContestPhase.kt
@@ -1,0 +1,7 @@
+package at.hannibal2.skyhanni.features.garden.contest
+
+enum class FarmingContestPhase {
+    START,
+    STOP,
+    CHANGE
+}